safari-534.53-branch/LayoutTests/platform/qt/fast/events/bogus-dropEffect-effectAllowed-expected.txt

This test checks that effectAllowed and dropEffect cannot be set to values that aren't defined in the spec

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.dataTransfer.effectAllowed is "all"
PASS event.dataTransfer.effectAllowed is "copy"
PASS event.dataTransfer.effectAllowed is "copy"
PASS event.dataTransfer.effectAllowed is "copyLink"
PASS event.dataTransfer.effectAllowed is "copyLink"
PASS event.dataTransfer.effectAllowed is "copyMove"
PASS event.dataTransfer.effectAllowed is "copyMove"
PASS event.dataTransfer.effectAllowed is "link"
PASS event.dataTransfer.effectAllowed is "linkMove"
PASS event.dataTransfer.effectAllowed is "move"
PASS event.dataTransfer.effectAllowed is "none"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.effectAllowed is "uninitialized"
PASS event.dataTransfer.effectAllowed is "uninitialized"

TEST COMPLETE
